requirements.txt
streamlit>=1.28.0
azure-ai-projects>=1.0.0b4
azure-identity>=1.15.0
azure-core>=1.29.0
